More updates based on PR feedback
SQUASH before merging

Made updates to test_xdd_debug_init.sh to just call XDD directly without
putting it in a string first.

Not sure what the idea was behind the redirection of the shell scripts
to the log file was. As fas as I can tell, it never resulted in anything
being written out to test-dir/logs. For now, I am just getting rid of
the redirection till someone can show me what it was suppose to do.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>