Fixing issue with test failures on CentOS
The following two tests would always fail on my CentOS 8 box:
1. test_xdd_passdelay
2. test_xdd_startdelay

The issue was that the time package was not installed. I updated the
build system to check if it is installed. If it is not installed, CMake
will state that both the tests test_xdd_passdelay and
test_xdd_startdelay will be skipped when running ctest. If it is
installed all tests will run like normal with the time command using the
proper path to the binary (/usr/bin/time). I also updated the main
README to state that time is an optional dependency.

I did a little bit of cleanup in the test directory. I got rid of the
test_config.example as that was used for the previous way to run the
tests for XDD. We now set those variables and create the test_config
file using the CMakeLists.txt in the test directory. I also got rid of
the the run_tests.sh script as that is no longer necessary as we moved
to CMake CTest now.

Finally, I updated the README file in the test directory to give a
description of how to properly add new tests using the CMake CTest
infrastructure.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>