Fixed issue with dryrun option that caused hanging
Fixed an issue with the dryrun option that would cause xdd to hang
whenever the option was enabled. The issue lied in src/client/results_manager.c,
as it was stuck waiting in a barrier for the other threads to also enter the
barrier, which they would never do as the other threads have already exited.

Also added a test to ensure that the dryrun option functions correctly.

Signed-off-by: MigeljanImeri <ImeriMigel@gmail.com>