Modified tests to no longer use relative pathing
Modified the way tests source the config and common.sh files, they are
now able to be called from anywhere as absolute pathing is now used.

Also added the test_xdd_createnewfiles.sh test to the list of tests to
run, as it was not being ran before.

Fixed an issue with test_xdd_dryrun that caused it not to be working
properly.

Signed-off-by: MigeljanImeri <ImeriMigel@gmail.com>