Updating how we file datapatterns work
Previously when the option -datapattern file/wholefile was used, XDD
would open the file for every target and keep it open till the run was
over. This is really unnecessary as the file only needs to be opened to
read the datapattern in to set the datapattern in the target. Once that
has happen, the file can immediately be closed. This is a good idea as
there are limits to how many files can be opened in Linux (ulimit). XDD
should not tie up the number of avialable open file descriptors just to
set a datapattern over an entire XDD run.

The code has been updated to open the file, set the target data_pattern
buffer, and then immediately close the file. I have added a new
functional test case test_xdd_file_datapattern.sh to test out that these
this modification did not break the expected functionality of
-datapattern file/wholefile.

There was also just some general cleanup of the code that was done as
part of this PR.

Signed-off-by: Brian Atkinson <batkinson@lanl.gov>