Simplify -seek Command and Reduce Required Arguments (#34)
This commit improves seek file handling by adding metadata
for the "Number of Requests" at the beginning of the seek file,
eliminating the need to read the entire file to retrieve this value.
The program now prioritizes the block size and request size specified
in the seek file, replacing any values provided via the command line
when the `-seek load` option is used.

Additionally, this commit fixes a segmentation fault issue that
occurred when no options were provided with the `-seek` command.
Comments in the `xdd_func` array have also been updated to match the
current function setup. A test script has been added to ensure the
program correctly sets and maintains the block size, request size,
and "Number of Requests" values during file loading.

Signed-off-by: Cade Gore <79346293+cadegore@users.noreply.github.com>