Fix datapattern parsing for multiple targets
When -datapattern file was passed for a xdd run with multiple
targets, the datapattern would only get set on the first target.
The rest of the targets would get an arbitrary buffer, leading to
inconsistent benchmarking results. This was because we were not
setting this while looping over multiple targets.

Signed-off-by: MigeljanImeri <imerimigel@gmail.com>